Allow erlexec to run without priv dir
When erlexec is run as part of an escript, we don't have access to the
priv dir, and calling `code:priv_dir` will return an error.  This
updates the default function of erlexec to support this case.

Users will need to manually provide the path to erlexec themselves in
these cases.